[rename] package name change member -> sample
