+++ src/main/java/com/study/jinyoung/domain/sample/dto/request/CreateSampleRequestDto.java
@@ -1,4 +1,4 @@
-package com.study.jinyoung.domain.member.dto.request;
+package com.study.jinyoung.domain.sample.dto.request;
 
 import lombok.Getter;
