+++ src/main/java/com/study/jinyoung/domain/sample/repository/SampleRepository.java
@@ -1,6 +1,6 @@
-package com.study.jinyoung.domain.member.repository;
+package com.study.jinyoung.domain.sample.repository;
 
-import com.study.jinyoung.domain.member.entity.Sample;
+import com.study.jinyoung.domain.sample.entity.Sample;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SampleRepository extends JpaRepository<Sample, Long> {
